Create Find files with given extension and remove the extension using ansible
I was searching for a process in the internet for renaming files in Linux using Ansible/removing file extensions from some files . Most of the posts talked about renaming to some other area, but I couldn't find much on removing the extension. After multiple tries I was able to achieve what is needed and I thought it is good the share the same . the same playbook with little modifications can be used for multiple purposes. 

The playbook is written in such a way that , Path of the files, required extension to remove and the hosts is taken as an input . 

https://linuxacatalyst.blogspot.com/